rc ref match fix

+++ src-cpp/index.h
@@ -209,7 +209,7 @@ typedef struct {
 // **** Read Set Index ****
 
 struct ref_match_t {
-	uint64_t pos;
+	uint32_t pos;
 	uint32 len;
 	bool rc;
 	ref_match_t(seq_t _pos, uint32 _len, bool _rc) : pos(_pos), len(_len), rc(_rc) {}
